Update the http proxy strategy
- Use the --proxy option as the default way to communicate with
  Orchestrate.io. You can either use --proxy https://api.orchestrate.io
every time you run the ember server command or the proxy path to your
.ember-cli config file.
- For apps that need to communicate to more than just the Orchestrate.io
  api, use the provided orchestrate-proxy generator to create a node
http-proxy to Orchestrate. You will also need to add '/orchestrate' as
the host in addition to your api key to the Orchestrate adapter. For
production apps, the host can be changed to point to your own proxy to
Orchestrate.