Fixes changes made in 2944675.
`loadMore` needs to be on the RecordArray class so that it is available
in both `AdapterPopulatedRecordArrays` and `ManyArrays`.
However, since a `findAll` request, which is called when the `find` method is
used without a query object passed in, doesn't set metadata. `hasNext` and
`loadMore` will only work when a query object is passed into both the
`find` and `findQuery` methods.